cleanup Gateway code
